refactor: NBT-69 entity파일들 경로 이동
