+++ src/main/java/com/newbit/purchase/command/domain/aggregate/ColumnPurchaseHistory.java
@@ -1,4 +1,4 @@
-package com.newbit.purchase.command.domain;
+package com.newbit.purchase.command.domain.aggregate;
 
 import jakarta.persistence.*;
 import lombok.Getter;
